backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1

numpy==2.0.2
pandas==2.2.2
matplotlib==3.10.0
seaborn==0.13.2

scikit-learn==1.6.1
xgboost==3.1.2
joblib==1.5.2
imbalanced-learn==0.14.0

mysql-connector-python==9.1.0
